httpx: update http code to support Go1.7 context
This change cleans up the existing httpserver code by replacing it with
the httpx (aka. http extensions) package.

The httpx package is much simplified by taking advantage of the addition
of golang.org/x/net/context into the Go standard library as part of the
golang-1.7 release.

The previous httpserver was a simplified/watered version of Gin-Gonic
targeting just JSON over HTTP.  The new httpx package is based off of
the very popular httprouter, but modifies it to use the Context for
path parameters, instead of a custom function prototype, allowing us to
use the builtin http.Handler type seamlessly w/ the speed of the radix
tree.  The new type is "httpx.Mux".

We've also defined a "httpx.Chain" as a simple and reliable tool to
wrap middleware around a handler.