httpx: fix panic race condition with closeNotify
The CloseNotify function creates a channel which MAY receive a signal
before ServeHTTP returns; but calling CloseNotify MUST occur before the
http handler returns.

This change moves the call to CloseNotify up out of the goroutine so
that there is no race condition and it is always called before the
handler returns.